refactor iocage.lib.Config.File.Prototype
